Refactor: Split generate-resume function into modules
Refactored the `supabase/functions/generate-resume/index.ts` file into smaller, more manageable modules to improve code organization and maintainability. This includes separating concerns such as prompt building, API interaction, and content sanitization. Unused imports and files were removed.
